refactor(website): extract shared discoverPageRoutes utility
Both app/sitemap.ts and app/sitemap.md/route.ts walked the app/ directory to discover page routes, with two near-identical copies of the recursion + exclusion set. Move that logic into utils/discover-page-routes.ts so the exclusion set (api, open-file, anything with a dot) is defined once and shared.

Reported by Cursor Bugbot.
