+++ apps/website/app/sitemap.ts
@@ -1,44 +1,12 @@
 import type { MetadataRoute } from "next";
-import { readdirSync, statSync } from "fs";
 import { join } from "path";
+import { discoverPageRoutes } from "@/utils/discover-page-routes";
 
 const BASE_URL = "https://react-grab.com";
 
-const EXCLUDED_PATHS = new Set(["api", "open-file"]);
-
-const getRoutes = (directory: string, basePath = ""): Array<string> => {
-  const routes: Array<string> = [];
-  const entries = readdirSync(directory);
-
-  for (const entry of entries) {
-    const fullPath = join(directory, entry);
-    const routePath = basePath ? `${basePath}/${entry}` : entry;
-
-    if (EXCLUDED_PATHS.has(entry)) {
-      continue;
-    }
-
-    const stat = statSync(fullPath);
-
-    if (stat.isDirectory()) {
-      const hasPage = readdirSync(fullPath).some(
-        (file) => file === "page.tsx" || file === "page.ts",
-      );
-
-      if (hasPage) {
-        routes.push(routePath);
-      }
-
-      routes.push(...getRoutes(fullPath, routePath));
-    }
-  }
-
-  return routes;
-};
-
 const sitemap = (): MetadataRoute.Sitemap => {
   const appDirectory = join(process.cwd(), "app");
-  const routes = getRoutes(appDirectory);
+  const routes = discoverPageRoutes(appDirectory);
 
   const sitemapEntries: MetadataRoute.Sitemap = [
     {
